Add EPOLLET flag when adding fd to epoll to support upstream driver
The upstream driver uses UIO and requires the fd that is added to
the epoll loop to be added in edge triggered mode or you will get
a notification every time you loop round the epoll. It should be
noted that this also applies to any applications that use external
polling mode who will also need to remember to use the EPOLLET flag
if using the driver fd's within an epoll loop. Adding EPOLLET is
backwards compatible with the existing driver.

Signed-off-by: Steve Linsell <stevenx.linsell@intel.com>