Refactor chained cipher implementation and Add support for openssl pipelining.
    Pipelining provides the capability for parallel encryption of multiple buffers.
All the buffers in pipeline are submitted as a batch to the accelerator. The operation
is complete when all the buffers have been processed successfully. A failure for a single buffer
results in a failure for the whole pipeline. Pipelines are always offloaded to the accelerator
ignoring small packet offload thresholds.

Change-Id: Ia914f74290465b458ca8b0601e61a4e3c6f6746d

Signed-off-by: Andrea Grandi <andrea.grandi@intel.com>