Fix for engine async notification fd getting closed too early
The issue was caused by a race condition between the polling thread and the submission thread.
The submission thread had been optimized to detect if the polling thread had completed pulling
the message response from the hardware before the need to switch out of the fibre. If this
happens then the fd was getting closed and removed from the ASYNC_WAIT_CTX. This works if
using a new fd but if using a prexisting fd then it is undesirable to close it here, it is
better to leave it alone and allow it to continue to be reused for the life of the connection.
This fix detects whether the fd is newly created and only closes and removes it from the
ASYNC_WAIT_CTX if that is the case.

Change-Id: I2e46b8e2922a653ff858bb93f28abd84572dfb71
Signed-off-by: Steve Linsell <stevenx.linsell@intel.com>