Add slot size optimisations within qae_mem_utils.c
Slot sizes were defined for example as 256 bytes but then the slot header
and alignment were taken off of this so the slot was actually smaller.
For the asymmetric crypto the key sizes tend towards regular allocations
of 128 or 256 bytes. This change takes account of the slot header and
alignment bytes when declaring the slots so for instance you will
end up with slots that really have 256 bytes of usable space. The idea is
to minimise the amount of wasted space.
This change also adds some extra slot sizes. This may result in additional
memory usage when using very few connections but under load should reduce
the overall memory footprint.

Change-Id: Iaac7e05931afc1b72b2e14d8dc511c08935e660a
Signed-off-by: Steve Linsell <stevenx.linsell@intel.com>