Remove non terminal errors from polling loops
Errors were being flagged on the OpenSSL error stack when a non
terminal error occurs when attempting to poll the rings for
responses. When an error occurs it could for instance be because
the hardware is resetting and will subsequently recover. As such
we do not quit the polling loop but continue and try again. If we
flag an error it should only be flagged once not every time round
the loop. In this case it is better to not flag the error at all.

Change-Id: I3e7d019c0526d311fbf3035f1ae3aed4485e0b0e
Signed-off-by: root <root@wgclpixa00388425.ir.intel.com>