Reinstate checks for memory driver and accelerator
The checks at startup for the memory driver and for
the acceleration device were not getting called when
running as a dynamic engine. As it is always run as
a dynamic engine it was impossible for these checks
to be invoked. They have now been moved to the
bind_qat() function instead.

Change-Id: Ia6972652c8e929415a623dfd53b2a0f7d882d86a
Signed-off-by: root <root@wgclpixa00388425.ir.intel.com>