Fix crash in signal code after forking.
Thanks to user i10a for reporting this issue.
After a fork the engine is not initialised in the child until the
first call of get_next_inst(). qat_use_signals() also relies on the
engine being intialised in order for the application to know it can
send a signal to the polling thread. This change ensures that
qat_use_signals() will also initialise the engine after a fork, if
necessary.
get_next_inst() and qat_use_signals() have also been optimised so
that they do not use locking in the case where the engine has already
been initialised, which is most of the time.

Change-Id: I5a1d088f7f6cbea8cef240a15a05ae2df263fa54
Signed-off-by: Steve Linsell <stevenx.linsell@intel.com>