Optimisation to reduce the number of signals sent.
Currently a signal is sent on every qat offload to resume the polling
thread if it has been suspended. This change uses thread local variables
to reduce the signals sent to one per thread when transitioning from
zero to one offload in flight for that thread.
Note we can't easily reduce it further to only when the overall offloads
in flight transition from zero to one as that would involve a big
performance penalty in synchronising between threads.

Change-Id: I81fdb3c32e00b9c2c3d81b7def44d999365d6707
Signed-off-by: Steve Linsell <stevenx.linsell@intel.com>