Add support for the async callback mechanism in the QAT Engine.
The async callback mechanism is only supported in the OpenSSL master
branch.  These changes will auto detect if building for the master
branch and enable the feature.  The feature will only be used if the
application sets the relevant callback. If the callback is not set or
if building against on older version of OpenSSL then the file descriptor
notification method will be used.

Change-Id: Iba22d521e7bf62e3ac84f75968c77228dff1b80e
Signed-off-by: Ping Yu <ping.yu@intel.com>
Signed-off-by: Steve Linsell <stevenx.linsell@intel.com>