configure: Fix configure.ac when setting qatlib --with-qat_hw_dir
When qatlib is installed in a non-standard location or on an OS
that is not Red Hat/Fedora, the location of the library is not
detected. Fix this by introducing a temporary variable containg
the value of --with-qat_hw_dir if given, or /usr if not. Check
for the header file using the temporary value and set
$with_qat_hw_dir once more to handle the case of the prefix not
being set. After this the logic in configure.ac follows the
same path as before, with the modification that $with_qat_hw_dir
is used with the library and include files.

Signed-off-by: Patrik Flykt <patrik.flykt@intel.com>
Signed-off-by: Yogaraj Alamenda <yogarajx.alamenda@intel.com>